Create addonconf module

+++ src/make-xpi.py
@@ -5,23 +5,14 @@
 # License, v. 2.0. If a copy of the MPL was not distributed with this
 # file, You can obtain one at http://mozilla.org/MPL/2.0/.
 
-"""
-Usage:
-
-./make-xpi.py
-./make-xpi.py theme
-./make-xpi.py extension
-./make-xpi.py clean
-"""
-
 import sys
 import os
 import shutil
-import json
 import argparse
 
 sys.path.insert(0, "./build")
 
+import addonconf
 from themebuilder import ThemeBuilder
 from extensionbuilder import ExtensionBuilder
 from packagebuilder import PackageBuilder
@@ -41,10 +32,30 @@ def main():
 
     action = args.action
 
-    #
-    # Clean up
-    #
+    # Create config = argparse + config.json
+    config = addonconf.load("config.json")
+    if not config:
+        sys.exit(1)
 
+    if "VERSION" in os.environ:
+        config["version"] = os.environ.get("VERSION")
+        config["override-version"] = True
+    if args.version:
+        config["version"] = args.version
+        config["override-version"] = True
+
+    if args.target_version:
+        config["target-version"] = args.target_version
+
+    config["verbose"] = 0
+    if args.verbose:
+        config["verbose"] = args.verbose
+
+    config = addonconf.validate(config, action)
+    if not config:
+        sys.exit(1)
+
+    # Clean up
     if action == "clean":
         if os.path.isdir(".build"):
             shutil.rmtree(".build")
@@ -55,56 +66,19 @@ def main():
                 os.remove(os.path.join("build", name))
         sys.exit(0)
 
-    #
-    # Create config = argparse + config.json
-    #
-
-    try:
-        with open("config.json", "r") as config_file:
-            config = json.load(config_file)
-    except FileNotFoundError:
-        print("%s: %s not found" % (sys.argv[0], "config.json"))
-        sys.exit(1)
-    except ValueError as e:
-        print("%s: parse error: %s" % (sys.argv[0], "config.json"))
-        print(e)
-        sys.exit(1)
-
-    if "VERSION" in os.environ:
-        config["version"] = os.environ.get("VERSION")
-        config["override-version"] = True
-    if args.version:
-        config["version"] = args.version
-        config["override-version"] = True
-    if args.target_version:
-        config["target-version"] = args.target_version
-
-    config["verbose"] = 0
-    if args.verbose:
-        config["verbose"] = args.verbose
-
-    #
     # Theme building
-    #
-
     if action in ["theme", "all"]:
         builder = ThemeBuilder(config)
         print(":: Starting build theme...")
         builder.build()
 
-    #
     # Extension building
-    #
-
     if action in ["extension", "all"]:
         builder = ExtensionBuilder(config)
         print(":: Starting build extension...")
         builder.build()
 
-    #
     # Package building
-    #
-
     if action == "all":
         builder = PackageBuilder(config)
         print(":: Starting make package...")
